# musicbox-backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "MusicBox 音乐盒后端 API 文档",
        "title": "MusicBox API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/songs": {
            "get": {
                "description": "支持按关键字搜索和分页排序",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歌曲管理"
                ],
                "summary": "获取歌曲列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "upload_at",
                        "description": "排序字段",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "排序方式",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/songs/export": {
            "post": {
                "description": "整理数据库中所有歌曲（含歌词、封面）并打包为 Zip 直接下载",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "歌曲管理"
                ],
                "summary": "一键下载所有歌曲包",
                "responses": {
                    "200": {
                        "description": "歌曲压缩包",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "没有可导出的歌曲",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "导出失败",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/songs/upload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "上传音频文件及其相关信息（需管理员权限）",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歌曲管理"
                ],
                "summary": "上传歌曲",
                "parameters": [
                    {
                        "type": "string",
                        "description": "歌曲标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "艺术家",
                        "name": "artist",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "专辑",
                        "name": "album",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "歌曲音频文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "封面图片",
                        "name": "cover",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "歌词文件",
                        "name": "lyric",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "上传成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Song"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/songs/{id}": {
            "get": {
                "description": "根据歌曲 ID 获取歌曲完整信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歌曲管理"
                ],
                "summary": "获取歌曲详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "歌曲 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Song"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的 ID",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "歌曲未找到",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改现有歌曲的信息或替换相关文件（需管理员权限）",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歌曲管理"
                ],
                "summary": "更新歌曲",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "歌曲 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "歌曲标题",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "艺术家",
                        "name": "artist",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "专辑",
                        "name": "album",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "新歌曲音频文件",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "新封面图片",
                        "name": "cover",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "新歌词文件",
                        "name": "lyric",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Song"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "歌曲未找到",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据 ID 删除歌曲记录及相关文件（需管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歌曲管理"
                ],
                "summary": "删除歌曲",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "歌曲 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "歌曲未找到",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Song": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string"
                },
                "artist": {
                    "type": "string"
                },
                "coverUrl": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "fileUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lyricUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uploadAt": {
                    "type": "string"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT Token，格式：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}